Wired up GDBRemoteCommunicationServer to fire off a NativeProcessLinux.
This build compiles but doesn't link:

* Need to implement NativeProcessLinux implementation of
  NativeProcessProtocol methods.

* Need to modify PlatformLinux::LaunchDebugProcess () to retrieve or be
  passed the real BroadcasterManager. Currently it uses NULL.

Broke out GDBRemoteCommunicationServer to call different launch methods
when it is lldb-platform vs. when it is lldb-gdbserver. When not the gdb
server, it does what it did before when launching a new process. When it
is the gdb server, it now asks the platform for a
NativeProcessProtocolSP.