Pushed Debugger instance as BroadcasterManager through to PlatformLinux within lldb-gdbserver.
Discovered that the Debugger instance is the implementer of
BroadcasterManager.  I now pass the DebuggerSP from the lldb-gdbserver
to the GDBRemoteCommunicationServer in the case of lldb-gdbserver.  I
added constructor asserts to ensure we always have a DebuggerSP when
constructing when we're not the lldb-platform since we'll need it.

Tweaked signatures for Platform::LaunchDebugProcess () to include the
BroadcasterManager pointer.

Compiles, still does not link due to the missing impl in
NativeProcessLinux for the NativeProcesProtocol methods.