Added NativeThreadLinux class.
Implemented the Read and Write register methods in terms of
NativeProcessLinux methods. Added some helpers to NativeProcessLinux to
support doing the lldb register number mapping to the Linux register
handling.

I'm ending up bringing in some RegisterContextPOSIX* class content into
NativeProcessLinux to support doing the register detail lookup from the
register_infos. I can't use the Context as it is currently written since
it depends on a Thread and Process. It might be possible/desireable to
refactor the RegisterContextPOSIX* classes to take the thread/process
info directly. For now I'm ignoring this.