Added $qProcessInfo support to llgs/lldb-platform.
It's not yet identical to debugserver's version, but it does contain the
critical pid info and almost everything else. Requires debugserver
concordance fixup on os version (ios) and vendor, and the ptrsize
key/value.

The $qC and $? thread equivalence test is now enabled and works on
llgs/Linux.

The $qProcessInfo test now works on llgs/Linux.