Fixed up llgs to detach from process when it exits.
Enabled the run-to-completion-after-attach test for llgs.

Handling of the 'k' packet now properly exits llgs.

A process exiting now handles exiting llgs.

Added an m_exit_now variable to GDBRemoteCommunicationServer
which gets checked in the packet handling loop.  If set, it
responds that llgs should exit.

Discovered that I cannot nuke the m_debugged_process_sp in
the process state change processing - that is occurring synchronously
within the NativeProcessProtocol signal handling innards which,
when coming out of process state change notification, still needs
to be live.  This will need to be decoupled later (hey - that's
the whole broadcast/receiver infrastructure we expected to have
earlier - may want to consider re-instating that).