Implemented $q{f,s}ThreadInfo for llgs.
Also fixed up incorrect impl of $qC - that was still responding with
the process id.  Added comments around it being the wrong response for
lldb-platform, but kept it as process id for that case.  I believe we
should change it to return $#00 for lldb-platform while llgs responses
with the current inferior thread id.